refactor: 将 disable cv warning 的宏额外封装了一层头文件
