+++ src/MeoAssistant/AbstractImageAnalyzer.h
@@ -1,10 +1,6 @@
 #pragma once
 
-#include "Config.hpp"
-ASST_SUPPRESS_CV_WARNINGS_START
-#include <opencv2/opencv.hpp>
-ASST_SUPPRESS_CV_WARNINGS_END
-
+#include "NoWarningCVMat.h"
 #include "AsstTypes.h"
 
 //#ifndef  ASST_DEBUG
