+++ src/MeoAssistant/AsstBattleDef.h
@@ -4,10 +4,7 @@
 #include <string>
 #include <unordered_map>
 
-#include "Config.hpp"
-ASST_SUPPRESS_CV_WARNINGS_START
-#include <opencv2/core/mat.hpp>
-ASST_SUPPRESS_CV_WARNINGS_END
+#include "NoWarningCVMat.h"
 
 #include "AsstTypes.h"
