+++ src/MeoAssistant/Controller.h
@@ -15,10 +15,7 @@
 #include <Windows.h>
 #endif
 
-#include "Config.hpp"
-ASST_SUPPRESS_CV_WARNINGS_START
-#include <opencv2/core/mat.hpp>
-ASST_SUPPRESS_CV_WARNINGS_END
+#include "NoWarningCVMat.h"
 
 #include "AsstTypes.h"
 #include "AsstMsg.h"
