+++ src/MeoAssistant/InfrastSmileyImageAnalyzer.cpp
@@ -1,5 +1,7 @@
 #include "InfrastSmileyImageAnalyzer.h"
 
+#include "NoWarningCV.h"
+
 #include "AsstUtils.hpp"
 #include "MultiMatchImageAnalyzer.h"
 #include "Resource.h"
