+++ src/MeoAssistant/MeoAssistant.vcxproj.filters
@@ -366,7 +366,13 @@
     <ClInclude Include="RoguelikeResetTaskPlugin.h">
       <Filter>头文件\Task\Plugin\Roguelike</Filter>
     </ClInclude>
-    <ClInclude Include="Config.hpp">
+    <ClInclude Include="AsstConf.h">
+      <Filter>头文件\Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="NoWarningCV.h">
+      <Filter>头文件\Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="NoWarningCVMat.h">
       <Filter>头文件\Utils</Filter>
     </ClInclude>
   </ItemGroup>
