+++ src/MeoAssistant/NoWarningCV.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "AsstConf.h"
+
+ASST_SUPPRESS_CV_WARNINGS_START
+#include <opencv2/opencv.hpp>
+ASST_SUPPRESS_CV_WARNINGS_END
