+++ src/MeoAssistant/NoWarningCVMat.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "AsstConf.h"
+
+ASST_SUPPRESS_CV_WARNINGS_START
+#include <opencv2/core/mat.hpp>
+ASST_SUPPRESS_CV_WARNINGS_END
