+++ src/MeoAssistant/OcrPack.cpp
@@ -2,11 +2,8 @@
 
 #include <filesystem>
 
+#include "NoWarningCV.h"
 #include <PaddleOCR/paddle_ocr.h>
-#include "Config.hpp"
-ASST_SUPPRESS_CV_WARNINGS_START
-#include <opencv2/opencv.hpp>
-ASST_SUPPRESS_CV_WARNINGS_END
 
 #include "AsstUtils.hpp"
 #include "Logger.hpp"
