+++ src/MeoAssistant/TilePack.cpp
@@ -1,6 +1,6 @@
 #include "TilePack.h"
 
-#include "Config.hpp"
+#include "AsstConf.h"
 ASST_SUPPRESS_CV_WARNINGS_START
 #include <Arknights-Tile-Pos/TileCalc.hpp>
 ASST_SUPPRESS_CV_WARNINGS_END
